Add files via upload
Losses in the banking and money lending industry are attributed mainly to the large number of defaulting customers. In turn this project looks at the various features that initiate this situation and out of it a machine learning model is formed. This model is effective as it'll minimizes the losses from defaulting. For it predicts the possibility of a customer to default from the data given.